Accented characters using Shift-Ctrl-X <accent-char><letter>
Easy to remember!
` -> grave
' -> acute
^ -> circumflex
~ -> tilde
: -> umlaut
, -> cedilla

I haven't implemented caron, mostly because I haven't had use for it and
don't offhand know of a suitable (non-alphabetic) character to signify
it. We could use doubled-'^' circumflex, but that would require a little
more logic, which also doesn't seem worth the while until we need it.