Migrate several app jumps to PowerToys keyboard Manager
YAY YAY YAY!

I have some more to figure out, but I've taken care of my most frequent
jumps and the ones that also most frequently inadvertently start MS Office:
Edge and Chrome.

Apparently a recent PowerToys release includes app-start functionality, and
that can be configured to jump to the window if the app is already
open. (In fact, that's the default choice.)

You need to specify the path to the executable, which fortunately Task
Manager makes easy (right-click menu "Properties" on the running app).

Want to figure out the right path for Ubuntu apps, and there are other apps
to migrate, but jumping to them doesn't tend to incidentally start MS
Office, so I can continue with this hybrid configuration.